refactoring: move shortcode helper functions to conversion and registry modules
